requirements.txt
fastapi
flask
numpy
openai
python-dotenv
pandas
pydantic
requests
python-multipart
uvicorn
gunicorn
matplotlib
groq

